# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ivmp-service API",
        "title": "ivmp-service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/swagger/ivmp-service",
    "paths": {
        "/ai-model": {
            "get": {
                "description": "query objects",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ai_model"
                ],
                "summary": "query objects",
                "parameters": [
                    {
                        "type": "string",
                        "description": "_select",
                        "name": "_select",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "order",
                        "name": "_order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "created_by",
                        "name": "created_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "created_time",
                        "name": "created_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "updated_by",
                        "name": "updated_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "updated_time",
                        "name": "updated_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "config",
                        "name": "config",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "description",
                        "name": "description",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "version",
                        "name": "version",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "file_ext",
                        "name": "file_ext",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sub_type",
                        "name": "sub_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects array",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.Ai_model"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "save",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ai_model"
                ],
                "summary": "save",
                "parameters": [
                    {
                        "description": "object",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Ai_model"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Ai_model"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/ai-model/batch-delete": {
            "post": {
                "description": "batch delete",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ai_model"
                ],
                "summary": "batch delete",
                "parameters": [
                    {
                        "description": "id array",
                        "name": "ids",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/ai-model/batch-upsert": {
            "post": {
                "description": "batch update",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ai_model"
                ],
                "summary": "batch update",
                "parameters": [
                    {
                        "description": "objects array",
                        "name": "entities",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/ai-model/groupby": {
            "get": {
                "description": "GroupBy, for example,  _select=level, then return  {level_val1:sum1,level_val2:sum2}, _where can input status=0",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ai_model"
                ],
                "summary": "GroupBy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "_select",
                        "name": "_select",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "_where",
                        "name": "_where",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects array",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "additionalProperties": true
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/ai-model/page": {
            "get": {
                "description": "page query, _page(from 1 begin), _page_size, _order, and others fields, status=1, name=$like.%CAM%",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ai_model"
                ],
                "summary": "page query",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "current page",
                        "name": "_page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "_page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "order",
                        "name": "_order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "created_by",
                        "name": "created_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "created_time",
                        "name": "created_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "updated_by",
                        "name": "updated_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "updated_time",
                        "name": "updated_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "config",
                        "name": "config",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "description",
                        "name": "description",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "version",
                        "name": "version",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "file_ext",
                        "name": "file_ext",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sub_type",
                        "name": "sub_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects array",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/common.Page"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "items": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/model.Ai_model"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/ai-model/{id}": {
            "delete": {
                "description": "delete",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ai_model"
                ],
                "summary": "delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "实例id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Ai_model"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.Page": {
            "type": "object",
            "properties": {
                "items": {},
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "common.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "model.Ai_model": {
            "type": "object",
            "properties": {
                "config": {
                    "description": "配置,json格式，每种类型不同的格式",
                    "type": "string"
                },
                "created_by": {
                    "description": "创建者",
                    "type": "string"
                },
                "created_time": {
                    "description": "创建时间",
                    "type": "string"
                },
                "description": {
                    "description": "描述",
                    "type": "string"
                },
                "file_ext": {
                    "description": "文件后缀",
                    "type": "string"
                },
                "id": {
                    "description": "id",
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "sub_type": {
                    "description": "sub_type",
                    "type": "string"
                },
                "type": {
                    "description": "type",
                    "type": "string"
                },
                "updated_by": {
                    "description": "更新者",
                    "type": "string"
                },
                "updated_time": {
                    "description": "更新时间",
                    "type": "string"
                },
                "version": {
                    "description": "版本",
                    "type": "string"
                }
            }
        },
...
    }
}